requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.0.2
djangorestframework==3.10.2
sqlparse==0.3.0
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.2
